requirements.txt
pyTelegramBotAPI
yt-dlp
requests
curl-cffi
Flask
gunicorn
python-dotenv